Refactor for improved structure
