+++ pkg/plugin/settings.go
@@ -4,6 +4,8 @@ import (
 	"encoding/json"
 
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
+	"github.com/grafana/sentry-datasource/pkg/errors"
+	"github.com/grafana/sentry-datasource/pkg/util"
 )
 
 type SentryConfig struct {
@@ -14,13 +16,13 @@ type SentryConfig struct {
 
 func (sc *SentryConfig) Validate() error {
 	if sc.URL == "" {
-		return ErrorInvalidSentryConfig
+		return errors.ErrorInvalidSentryConfig
 	}
 	if sc.OrgSlug == "" {
-		return ErrorInvalidOrganizationSlug
+		return errors.ErrorInvalidOrganizationSlug
 	}
 	if sc.authToken == "" {
-		return ErrorInvalidAuthToken
+		return errors.ErrorInvalidAuthToken
 	}
 	return nil
 }
@@ -28,22 +30,22 @@ func (sc *SentryConfig) Validate() error {
 func GetSettings(s backend.DataSourceInstanceSettings) (*SentryConfig, error) {
 	config := &SentryConfig{}
 	if err := json.Unmarshal(s.JSONData, config); err != nil {
-		backend.Logger.Error(ErrorUnmarshalingSettings.Error())
-		return nil, ErrorUnmarshalingSettings
+		backend.Logger.Error(errors.ErrorUnmarshalingSettings.Error())
+		return nil, errors.ErrorUnmarshalingSettings
 	}
 	if config.URL == "" {
-		backend.Logger.Info("applying default sentry URL", "sentry url", DefaultSentryURL)
-		config.URL = DefaultSentryURL
+		backend.Logger.Info("applying default sentry URL", "sentry url", util.DefaultSentryURL)
+		config.URL = util.DefaultSentryURL
 	}
 	if config.OrgSlug == "" {
-		return nil, ErrorInvalidOrganizationSlug
+		return nil, errors.ErrorInvalidOrganizationSlug
 	}
 	if authToken, ok := s.DecryptedSecureJSONData["authToken"]; ok {
 		config.authToken = authToken
 	}
 	if config.authToken == "" {
-		backend.Logger.Error(ErrorInvalidAuthToken.Error())
-		return nil, ErrorInvalidAuthToken
+		backend.Logger.Error(errors.ErrorInvalidAuthToken.Error())
+		return nil, errors.ErrorInvalidAuthToken
 	}
 	return config, config.Validate()
 }
